Create trial design report

This function creates an html report describing the trial design created by the user with assign_rates() and includes figures showing machine alignment

Usage

make_trial_report(td, folder_path, trial_name = NA, keep_rmd = FALSE)

Arguments

td

trial design created by assign_rates()

folder_path

(character) path to the folder in which the report will be saved

trial_name

(character) name of trial to be used in report

keep_rmd

(logical) If FALSE (Default), the original rmd file will be deleted upon creating an html report. Otherwise, the rmd file will be saved in the folder specified by `folder_path`.

Value

path to the resulting html file (invisible)

Examples

#--- load experiment made by assign_rates() ---#
# \donttest{
data(td_single_input)
make_trial_report(
  td = td_single_input,
  folder_path = tempdir()
)
#> Not in lat/long. Returning original object.
# }