Prepare plot information for a single-input experiment (length in meter)

Prepare plot information for a single-input experiment case. All the length values need to be specified in meter.

Usage

prep_plot(
  input_name,
  unit_system,
  machine_width,
  section_num,
  harvester_width,
  plot_width = NA,
  headland_length = NA,
  side_length = NA,
  max_plot_width = NA,
  min_plot_length = NA,
  max_plot_length = NA
)

Arguments

input_name

(character) Input name

unit_system

(character) A character of either 'metric' or 'imperial' indicating the system of measurement used

machine_width

(numeric) A numeric number in units specified in unit_system that indicates the width of the applicator or planter of the input

section_num

(numeric) A numeric number that indicates the number of sections of the applicator or planter of the input

harvester_width

(numeric) A numeric number that indicates the width of the harvester

plot_width

(numeric) Default is c(NA, NA).

headland_length

(numeric) A numeric number that indicates the length of the headland (how long the non-experimental space is in the direction machines drive). Default is NA.

side_length

(numeric) A numeric number that indicates the length of the two sides of the field (how long the non-experimental space is in the direction perpendicular to the direction of machines). Default is NA.

max_plot_width

(numeric) Maximum width of the plots. Default is 36.576 meter (120 feet).

min_plot_length

(numeric) Minimum length of the plots. Default is 73.152 meter (240 feet).

max_plot_length

(numeric) Maximum length of the plots. Default is 91.440 meter (300 feet)

Value

a tibble with plot information necessary to create experiment plots

Examples

input_name <- "seed"
unit_system <- "metric"
machine_width <- 12
section_num <- 12
plot_width <- NA
harvester_width <- 24
prep_plot(input_name, unit_system, machine_width, section_num, harvester_width)
#> Since plot width was not provided via the `plot_with` option`, it was set to a least common multiplier of the widths of the machines.
#> # A tibble: 1 × 11
#>   input_name unit_system machine_width section_num section_width harvester_width
#>   <chr>      <chr>               <dbl>       <dbl>         <dbl>           <dbl>
#> 1 seed       metric                 12          12             1              24
#> # ℹ 5 more variables: plot_width <dbl>, headland_length <dbl>,
#> #   side_length <dbl>, min_plot_length <dbl>, max_plot_length <dbl>