Prepare plot information for a single-input experiment (feet)

Prepare plot information for a single-input experiment case. All the length values need to be specified in feet.

prep_plot_fs(
  input_name,
  machine_width,
  section_num,
  harvester_width,
  plot_width = NA,
  headland_length = NA,
  side_length = NA,
  max_plot_width = 120,
  min_plot_length = 240,
  max_plot_length = 260
)

Arguments

input_name

(character) Input name

machine_width

(numeric) A numeric number in feet that indicates the width of the applicator or planter of the input

section_num

(numeric) A numeric number that indicates the number of sections of the applicator or planter of the input

harvester_width

(numeric) A numeric number in feet that indicates the width of the harvester

plot_width

(numeric) Default is c(NA, NA).

headland_length

(numeric) A numeric number in feet that indicates the length of the headland (how long the non-experimental space is in the direction machines drive). Default is NA.

side_length

(numeric) A numeric number in feet that indicates the length of the two sides of the field (how long the non-experimental space is in the direction perpendicular to the direction of machines). Default is NA.

max_plot_width

(numeric) Maximum width of the plots in feet. Default is (36.576 meter).

min_plot_length

(numeric) Minimum length of the plots in feet. Default is 240 feet (73.152 meter).

max_plot_length

(numeric) Maximum length of the plots in feet. Default is 260 feet (79.248 meter).

Value

a tibble with plot information necessary to create experiment plots

Examples

input_name <- "seed"
machine_width <- 60
section_num <- 24
plot_width <- NA
harvester_width <- 30
prep_plot_fs(input_name, machine_width, section_num, harvester_width)
#>   input_name machine_width section_num section_width harvester_width plot_width
#> 1       seed        18.288          24         0.762           9.144      9.144
#>   headland_length side_length min_plot_length max_plot_length
#> 1          36.576       9.144          73.152          79.248