Prepare plot information for a two-input experiment (meter)

Prepare plot information for a two-input experiment case. All the length values need to be specified in meter.

prep_plot_md(
  form,
  machine_width,
  section_num,
  harvester_width,
  plot_width = c(NA, NA),
  headland_length = NA,
  side_length = NA,
  max_plot_width = measurements::conv_unit(120, "ft", "m"),
  min_plot_length = measurements::conv_unit(240, "ft", "m"),
  max_plot_length = measurements::conv_unit(260, "ft", "m")
)

Arguments

form

(vector of two elements)

machine_width

(vector of two elements)

section_num

(vector of two elements)

harvester_width

(numeric)

plot_width

(vector of two elements) Default is c(NA, NA)

headland_length

(numeric) Default is NA.

side_length

(numeric) Default is NA.

max_plot_width

(numeric) Default is 36.576 meter (120 feet).

min_plot_length

(numeric) Default is 79.248 meter (260 feet).

max_plot_length

(numeric) Default is 73.152 meter (240 feet).

Value

a tibble with plot information necessary to create experiment plots

Examples

form <- c("seed", "NH3")
machine_width <- c(12, 9)
section_num <- c(12, 1)
plot_width <- c(9.5, 36)
harvester_width <- 12
prep_plot_md(form, machine_width, section_num, harvester_width, plot_width)
#> For seed, the plot width you specified would cause mixed treatment problems. However, there is a plot width that avoids them. It is suggested that you use 12 as the plot width.
#> # A tibble: 2 × 11
#>   form  machin…¹ secti…² secti…³ harve…⁴ plot_…⁵ propo…⁶ headl…⁷ side_…⁸ min_p…⁹
#>   <chr>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 seed        12      12       1      12     9.5      12      24    9.14    73.2
#> 2 NH3          9       1       9      12    36        36      24    9.14    73.2
#> # … with 1 more variable: max_plot_length <dbl>, and abbreviated variable names
#> #   ¹​machine_width, ²​section_num, ³​section_width, ⁴​harvester_width,
#> #   ⁵​plot_width, ⁶​proposed_plot_width, ⁷​headland_length, ⁸​side_length,
#> #   ⁹​min_plot_length