Create data of input rate information for a single input

Create data of input rate information for a single input. This can be used to assign rates to experiment plots using assign_rates()

prep_rates_s(
  plot_info,
  gc_rate,
  unit,
  rates = NULL,
  min_rate = NA,
  max_rate = NA,
  num_rates = 5,
  design_type = NA,
  rank_seq_ws = NULL,
  rank_seq_as = NULL
)

Arguments

plot_info

(data.frame) plot information created by make_input_plot_data

gc_rate

(numeric) Input ate the grower would have chosen if not running an experiment. This rate is assigned to the non-experiment part of the field. This rate also becomes one of the trial input rates unless you specify the trial rates directly using rates argument

unit

(string) unit of input

rates

(numeric vector) Default is NULL. Sequence of trial rates in the ascending order.

min_rate

(numeric) minimum input rate. Ignored if rates are specified.

max_rate

(numeric) maximum input rate. Ignored if rates are specified

num_rates

(numeric) Default is 5. It has to be an even number if design_type is "ejca". Ignored if rates are specified.

design_type

(string) type of trial design. available options are Latin Square ("ls"), Strip ("strip"), Randomized Block ("rb"), Jump-conscious Latin Square ("jcls"), Sparse ("sparse"), and Extra Jump-conscious Alternate "ejca". See for more details.

rank_seq_ws

(integer) vector of integers indicating the order of the ranking of the rates, which will be repeated "within" a strip.

rank_seq_as

(integer) vector of integers indicating the order of the ranking of the rates, which will be repeated "across" strip for their first plots.

Value

data.frame of input rate information

Examples

plot_info <-
  prep_plot_fs(
    input_name = "seed",
    machine_width = 60,
    section_num = 24,
    harvester_width = 30,
    plot_width = 30
  )
#> 

prep_rates_s(
  plot_info,
  gc_rate = 30000,
  unit = "seeds",
  rates = c(20000, 25000, 30000, 35000, 40000)
)
#>   input_name design_type gc_rate  unit
#> 1       seed        jcls   30000 seeds
#>                                         rates_data rank_seq_ws rank_seq_as
#> 1 20000, 25000, 30000, 35000, 40000, 1, 2, 3, 4, 5        NULL        NULL