Trial design (single-input)

Trial design data created by assigning rates to experiment plots running the `assign_rates()` function. This data exists only for the purpose of making examples in some function references succinct.

Usage

td_single_input

Format

tbl_df tbl data.frame `td_single_input` A data frame with 1 rows and 9 columns:

input_name

input name

input_type

shorthand for the type of the input: "N" for nitrogen, "S" for seed, etc.

trial_design

experiment plots with input rats assigned as an sf object

design_type

type of the trial design used

unit

unit of the input

abline_type

(character) one of "free", "lock", "none" indicating the way ab-line is (or not) created

ab_lines

ab-lines for the applicator/planter as an sf object

harvest_ab_lines

ab-lines for the harvester as an sf object

field_sf

field boundary as an sf object

harvest_width

width of the harvester