Write trial design files for field implementation

Desc

write_trial_files(
  trial_design,
  folder_path = NA,
  tr_file_name = NA,
  a_abline_file_name = NA,
  h_abline_file_name = NA,
  overwrite = FALSE,
  zip = FALSE
)

Arguments

trial_design

(tibble) tibble that was created by applying assign_rate() to experimental plots made by make_exp_plots().

folder_path

(character) path to the folder in which the files will be saved

tr_file_name

(character) file name given to the trial design shape files

a_abline_file_name

(character) file name given to the applicator/planter ab-line shape files

h_abline_file_name

(character) file name given to the harvester ab-line shape files

overwrite

(logical) Default = FALSE. If TRUE, the existing shape file with the same file name will be overwritten.

zip

(logical) Default = FALSE. If TRUE, all the files that are being written will be zipped.

Value

write out files